QPID-5504 : refactring of queues, and introduce management node and amqp-management module
git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk/qpid@1567026 13f79535-47bb-0310-9956-ffa450edef68

+++ java/common/src/main/java/org/apache/qpid/framing/BasicContentHeaderProperties.java
@@ -30,8 +30,8 @@ import java.io.IOException;
 public class BasicContentHeaderProperties
 {
     //persistent & non-persistent constants, values as per JMS DeliveryMode
-    public static final int NON_PERSISTENT = 1;
-    public static final int PERSISTENT = 2;
+    public static final byte NON_PERSISTENT = 1;
+    public static final byte PERSISTENT = 2;
 
     private static final Logger _logger = LoggerFactory.getLogger(BasicContentHeaderProperties.class);
